+++ ImperatorToCk3.sln
@@ -18,11 +18,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "zlib", "ZipLib\extlibs\zlib
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ImperatorToCK3Tests", "ImperatorToCK3Tests\ImperatorToCK3Tests.vcxproj", "{2252FE41-5CF1-4585-9EED-10087D112F1F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frontend.Core", "frontend\Frontend.Core\Frontend.Core.csproj", "{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frontend.Core.Tests", "frontend\Frontend.Core.Tests\Frontend.Core.Tests.csproj", "{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frontend.Client", "frontend\ParadoxConverters.Frontend\Frontend.Client.csproj", "{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Fronter", "Fronter\Fronter.vcxproj", "{D4E43E3D-E131-466D-9AA7-704C2298EACC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -120,54 +116,20 @@ Global
 		{2252FE41-5CF1-4585-9EED-10087D112F1F}.Release|x64.Build.0 = Release|x64
 		{2252FE41-5CF1-4585-9EED-10087D112F1F}.Release|x86.ActiveCfg = Release|Win32
 		{2252FE41-5CF1-4585-9EED-10087D112F1F}.Release|x86.Build.0 = Release|Win32
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Debug|Win32.Build.0 = Debug|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Debug|x64.Build.0 = Debug|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Debug|x86.Build.0 = Debug|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Release|Win32.ActiveCfg = Release|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Release|Win32.Build.0 = Release|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Release|x64.ActiveCfg = Release|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Release|x64.Build.0 = Release|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Release|x86.ActiveCfg = Release|Any CPU
-		{94887089-4A35-40BE-BA3F-6C9F15AA2BDA}.Release|x86.Build.0 = Release|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Debug|Win32.Build.0 = Debug|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Debug|x64.Build.0 = Debug|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Debug|x86.Build.0 = Debug|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Release|Win32.ActiveCfg = Release|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Release|Win32.Build.0 = Release|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Release|x64.ActiveCfg = Release|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Release|x64.Build.0 = Release|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Release|x86.ActiveCfg = Release|Any CPU
-		{01F3A84A-BB75-4A4C-A34A-BC10F09E252C}.Release|x86.Build.0 = Release|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Debug|Win32.Build.0 = Debug|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Debug|x64.Build.0 = Debug|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Debug|x86.Build.0 = Debug|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Release|Win32.ActiveCfg = Release|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Release|Win32.Build.0 = Release|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Release|x64.ActiveCfg = Release|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Release|x64.Build.0 = Release|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Release|x86.ActiveCfg = Release|Any CPU
-		{0918DC8B-8D29-4B01-B83E-0896ECDF88A6}.Release|x86.Build.0 = Release|Any CPU
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Debug|Any CPU.ActiveCfg = Debug|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Debug|Win32.ActiveCfg = Debug|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Debug|Win32.Build.0 = Debug|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Debug|x64.ActiveCfg = Debug|x64
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Debug|x64.Build.0 = Debug|x64
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Debug|x86.ActiveCfg = Debug|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Debug|x86.Build.0 = Debug|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Release|Any CPU.ActiveCfg = Release|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Release|Win32.ActiveCfg = Release|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Release|Win32.Build.0 = Release|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Release|x64.ActiveCfg = Release|x64
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Release|x64.Build.0 = Release|x64
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Release|x86.ActiveCfg = Release|Win32
+		{D4E43E3D-E131-466D-9AA7-704C2298EACC}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
