+++ ImperatorToCk3/Copy_Files.bat
@@ -1,11 +1,12 @@
 echo on
 rem Copy converter data files
-copy "Data_Files\configuration.txt" "..\Release\ImperatorToCK3\configuration.txt"
 rem copy "Data_Files\ReadMe.txt" "..\Release\ImperatorToCK3\readme.txt"
 rem copy "Data_Files\ReadMe.txt" "..\Release\readme.txt"
 rem copy "Data_Files\FAQ.txt" "..\Release\ImperatorToCK3\FAQ.txt"
-copy "Data_Files\ImperatorToCK3DefaultConfiguration.xml" "..\Release\Configuration\ImperatorToCK3DefaultConfiguration.xml"
-copy "Data_Files\SupportedConvertersDefault.xml" "..\Release\Configuration\SupportedConvertersDefault.xml"
+
+mkdir "..\Release\Configuration"
+copy "Data_Files\fronter-configuration.txt" "..\Release\Configuration\fronter-configuration.txt"
+copy "Data_Files\fronter-options.txt" "..\Release\Configuration\fronter-options.txt"
 
 rem Create Configurables
 del "..\Release\ImperatorToCK3\configurables" /Q
