--- ImperatorToCk3/Data_Files/ImperatorToCK3DefaultConfiguration.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<configuration Comment="This file is used to build the configuration parth of the frontend. This is done at run-time, so errors here might prevent the frontend from running.">
-<!--This part contains the various options that will be displayed as preferences in the converter ui, and outputted as configuration.txt as input for the converter tool itself-->
-<converterOptions>
-	<categories>
-		<category>
-			<friendlyName>Configuration</friendlyName>
-			<preferences>
-				<preference>
-					<name>ImperatorDeJure</name>
-					<friendlyName>Imperator de iure</friendlyName>
-					<description>Whether or not to base CK3 de iure on Imperator setup.</description>
-					<entryOptions>
-						<entryOption>
-							<name>1</name>
-							<friendlyName>Provinces and regions</friendlyName>
-							<description>Make Imperator provinces de iure CK3 duchies, make Imperator regions de iure CK3 kingdoms</description>
-							<isDefault>false</isDefault>
-						</entryOption>
-						<entryOption>
-							<name>2</name>
-							<friendlyName>Countries</friendlyName>
-							<description>Base CK3 de iure on Imperator countries</description>
-							<isDefault>false</isDefault>
-						</entryOption>
-						<entryOption>
-							<name>3</name>
-							<friendlyName>No</friendlyName>
-							<description>Keep vanilla CK3 de iure</description>
-							<isDefault>true</isDefault>
-						</entryOption>
-					</entryOptions>
-				</preference>
-				<preference>
-					<name>output_name</name>
-					<friendlyName>Mod Output Name (optional):</friendlyName>
-					<description>Please use ASCII latin characters only.</description>
-					<hasDirectlyEditableValue>true</hasDirectlyEditableValue>
-					<value></value>
-				</preference>
-			</preferences>
-		</category>
-	</categories>
-</converterOptions>
-</configuration>
