--- ImperatorToCk3/Data_Files/SupportedConvertersDefault.xml
@@ -1,148 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-Explanation: TODO:
-
-Topics to explain:
-
-tag vs internalTag: 
-* "tag" is for things that should be exported to the final configuration.txt file. It is essentially the name of a property in configuration.txt - examples being "CK2ExportDirectory", "V2Gametype", etc. 
-* "internalTag" is for tags the frontend needs, but which aren't part of configuration.txt. Still, I needed some way to find these in a generic way from code, which I can using internalTag. 
-  Example usage can be: "targetGameModPath". This lets me find the target game save game path, for instance, regardless of what the target game is. 
-
-
--->
-
-
-<configuration>
-	<gameConfigurationFile>Configuration\GameConfigurations.xml</gameConfigurationFile>
-	<supportedConverters>
-		<converter>
-			<name>ImperatorToCK3</name>
-			<friendlyName>From Imperator:Rome to Crusader Kings III</friendlyName>
-			<defaultConfigurationFile>Configuration\ImperatorToCK3DefaultConfiguration.xml</defaultConfigurationFile>
-			<userConfigurationFile></userConfigurationFile>
-			<sourceGame>ImperatorRome</sourceGame>
-			<targetGame>CrusaderKings3</targetGame>
-			<isEnabled>true</isEnabled>
-			<requiredItems>
-			
-				<requiredFile>
-					<tag>faq</tag>
-					<internalTag>faq</internalTag>
-					<friendlyName>faq</friendlyName>
-					<description>The faq .txt</description>
-					<isMandatory>false</isMandatory>
-					<hidden>true</hidden>
-					<alternatives>
-						<alternative>
-							<defaultLocationType>ConverterFolder</defaultLocationType>
-							<subFolderLocation>ImperatorToCK3</subFolderLocation>
-						</alternative>
-					</alternatives>
-					<predefinedFileName>FAQ.txt</predefinedFileName>
-				</requiredFile>
-
-				<requiredFile>
-					<tag>log</tag>
-					<internalTag>log</internalTag>
-					<friendlyName>log</friendlyName>
-					<description>The log .txt</description>
-					<isMandatory>false</isMandatory>
-					<hidden>true</hidden>
-					<alternatives>
-						<alternative>
-							<defaultLocationType>ConverterFolder</defaultLocationType>
-							<subFolderLocation>ImperatorToCK3</subFolderLocation>
-						</alternative>
-					</alternatives>
-					<predefinedFileName>log.txt</predefinedFileName>
-				</requiredFile>
-
-				<requiredFolder>
-					<tag>ImperatorDirectory</tag>
-					<friendlyName>Imperator: Rome Install directory</friendlyName>
-					<description>A path on your computer where Imperator: Rome is installed</description>
-					<isMandatory>true</isMandatory>
-					<alternatives>
-						<alternative>
-							<defaultLocationType>SteamFolder</defaultLocationType>
-							<autoDetectFromSteamId>859580</autoDetectFromSteamId> 
-						</alternative>
-					</alternatives>
-				</requiredFolder>
-
-				<requiredFolder>
-					<tag>ImperatorModsDirectory</tag>
-					<friendlyName>Imperator: Rome Mods Directory</friendlyName>
-					<description>A path on your computer where Imperator: Rome keeps mods</description>
-					<isMandatory>true</isMandatory>
-					<alternatives>
-						<alternative>
-							<defaultLocationType>WindowsUsersFolder</defaultLocationType>
-							<subFolderLocation>Paradox Interactive\Imperator\mod</subFolderLocation>
-						</alternative>
-					</alternatives>
-				</requiredFolder>
-
-				<requiredFolder>
-					<tag>CK3directory</tag>
-					<friendlyName>Crusader Kings III directory</friendlyName>
-					<description>A path on your computer where Crusader Kings III is installed</description>
-					<isMandatory>true</isMandatory>
-					<alternatives>
-						<alternative>
-							<defaultLocationType>SteamFolder</defaultLocationType>
-							<autoDetectFromSteamId>1158310</autoDetectFromSteamId> 
-						</alternative>
-					</alternatives>
-				</requiredFolder>
-
-				<requiredFolder>
-					<internalTag>targetGameModPath</internalTag>
-					<friendlyName>CK3 mods directory</friendlyName>
-					<description>A path on your computer where CK3 keeps mods</description>
-					<isMandatory>true</isMandatory>
-					<alternatives>
-						<alternative>
-							<defaultLocationType>WindowsUsersFolder</defaultLocationType>
-							<subFolderLocation>Paradox Interactive\Crusader Kings III\mod</subFolderLocation>
-						</alternative>
-					</alternatives>
-				</requiredFolder>
-
-				<requiredFile>
-					<tag>ConverterExe</tag>
-					<internalTag>Converterexe</internalTag>
-					<friendlyName>Converter .exe</friendlyName>
-					<description>The converter command line tool</description>
-					<isMandatory>true</isMandatory>
-					<extension>.exe</extension>
-					<alternatives>
-						<alternative>
-							<defaultLocationType>ConverterFolder</defaultLocationType>
-							<subFolderLocation>ImperatorToCK3</subFolderLocation>
-						</alternative>
-					</alternatives>
-					<extension>*.exe</extension>
-					<predefinedFileName>ImperatorToCK3Converter.exe</predefinedFileName>
-				</requiredFile>
-
-				<requiredFile>
-					<tag>SaveGame</tag>
-					<friendlyName>Savegame</friendlyName>
-					<description>The savegame to convert</description>
-					<extension>.rome</extension>
-					<isMandatory>true</isMandatory>
-					<alternatives>
-						<alternative>
-							<defaultLocationType>WindowsUsersFolder</defaultLocationType>
-							<subFolderLocation>Paradox Interactive\Imperator\save games</subFolderLocation>
-						</alternative>
-					</alternatives>
-				</requiredFile>
-
-			</requiredItems>
-		</converter>
-	</supportedConverters>
-</configuration>
